Refactor application into different classes to enforce SRP
